+++ R/selector.R
@@ -1,9 +1,11 @@
-new_selector <- function(css,
-                         xpath,
-                         id,
-                         class_name,
-                         name,
-                         filter = list(1)) {
+new_selector <- function(css = NULL,
+                         xpath = NULL,
+                         id = NULL,
+                         class_name = NULL,
+                         name = NULL,
+                         filter = list(1),
+                         multiple = FALSE,
+                         inner_multiple = NULL) {
   args <- list(
     css = css,
     xpath = xpath,
@@ -11,11 +13,13 @@ new_selector <- function(css,
     class_name = class_name,
     name = name,
     filter = filter,
-    to_be_filtered = length(filter)
+    to_be_filtered = length(filter),
+    multiple = multiple,
+    inner_multiple = inner_multiple
   )
 
   args <- args[!vapply(args, is.null, logical(1))]
-  args_without_filter <- args[!names(args) %in% c("filter", "to_be_filtered")]
+  args_without_filter <- args[!names(args) %in% c("filter", "to_be_filtered", "multiple")]
 
   if (length(args_without_filter) == 0) {
     stop_bad_selector()
@@ -55,16 +59,25 @@ use_selector <- function(selector, element, driver) {
     )
     return(elem_unique(elements, driver = driver))
   } else if (inherits(selector, "selenider_flatmap_selector")) {
-    elements <- selector$element
-    elements$driver <- driver
-    actual_elements <- get_elements(elements)
-    return(lazy_flatten(lazy_map(actual_elements, function(x) {
-      get_elements(list(
-        selectors = selector$selectors,
+    inner_multiple <- selector$inner_multiple
+    selector$multiple <- selector$inner_multiple
+    selector$inner_multiple <- NULL
+    selector$filter <- list()
+    selector$to_be_filtered <- 0
+    class(selector) <- class(selector)[class(selector) != "selenider_flatmap_selector"]
+    return(lazy_flatten(lazy_map(element, function(x) {
+      result <- get_elements(list(
+        selectors = list(selector),
         element = x,
         driver = driver,
-        to_be_found = length(selector$selectors)
+        to_be_found = 1
       ))
+
+      if (!inner_multiple) {
+        get_items(result, 1)
+      } else {
+        result
+      }
     })))
   }
 
@@ -72,10 +85,11 @@ use_selector <- function(selector, element, driver) {
 
   selector$filter <- NULL
   selector$to_be_filtered <- NULL
+  selector$multiple <- NULL
 
   if (length(filter) == 1 &&
-        identical(filter[[1]], 1) &&
-        length(selector) == 1) {
+    identical(filter[[1]], 1) &&
+    length(selector) == 1) {
     # If we are getting the first element, both CDP and Selenium have a
     # shortcut for this
     using <- switch(names(selector),